+++ components/binding/cmd/manager/main.go
@@ -1,9 +1,9 @@
 package main
 
 import (
-	"context"
 	"github.com/kyma-project/kyma/components/binding/internal/controller"
 	"github.com/kyma-project/kyma/components/binding/internal/storage"
+	"github.com/kyma-project/kyma/components/binding/internal/syncer"
 	"github.com/kyma-project/kyma/components/binding/internal/target"
 	"github.com/kyma-project/kyma/components/binding/internal/webhook"
 	bindingMutate "github.com/kyma-project/kyma/components/binding/internal/webhook/binding/mutate"
@@ -12,17 +12,14 @@ import (
 	targetKindValidate "github.com/kyma-project/kyma/components/binding/internal/webhook/targetkind/validate"
 	"github.com/kyma-project/kyma/components/binding/internal/worker"
 	bindingsv1alpha1 "github.com/kyma-project/kyma/components/binding/pkg/apis/v1alpha1"
-	"github.com/pkg/errors"
-	"k8s.io/client-go/dynamic"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sort"
-
 	log "github.com/sirupsen/logrus"
 	"github.com/vrischmann/envconfig"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/client-go/dynamic"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	k8sWebhook "sigs.k8s.io/controller-runtime/pkg/webhook"
 )
 
@@ -87,7 +84,7 @@ func main() {
 
 	targetKindWorker := worker.NewTargetKindWorker(targetKindStorage, dc)
 
-	err = syncTargetKinds(cli, targetKindWorker, logger)
+	err = syncer.NewExecutor(cli, logger).TargetKinds(targetKindWorker)
 	fatalOnError(err, "while syncing TargetKinds")
 
 	targetKindManager := target.NewHandler(dc, targetKindStorage)
@@ -104,29 +101,4 @@ func fatalOnError(err error, msg string) {
 	if err != nil {
 		log.Fatalf("%s: %s", msg, err.Error())
 	}
-}
-
-// syncTargetKinds assures that existing TargetKinds will be loaded before reconcilers are started
-func syncTargetKinds(cli client.Client, worker *worker.TargetKindWorker, log log.FieldLogger) error {
-	log.Info("Starting syncing TargetKinds on startup")
-
-	tks := &bindingsv1alpha1.TargetKindList{}
-	err := cli.List(context.Background(), tks)
-	if err != nil {
-		return errors.Wrap(err, "while listing TargetKinds")
-	}
-
-	tkList := tks.Items
-	sort.Slice(tkList, func(i, j int) bool {
-		return tkList[i].CreationTimestamp.Before(&tkList[j].CreationTimestamp)
-	})
-
-	for _, tk := range tkList {
-		_, err := worker.Process(&tk, log)
-		if err != nil {
-			return errors.Wrapf(err, "while processing TargetKind %s/%s", tk.Name, tk.Namespace)
-		}
-	}
-
-	return nil
 }
