+++ components/binding/internal/storage/storage.go
@@ -2,6 +2,7 @@ package storage
 
 import (
 	"fmt"
+	bindErr "github.com/kyma-project/kyma/components/binding/internal/errors"
 	"strings"
 	"sync"
 
@@ -55,7 +56,7 @@ func (s *storage) Get(kind v1alpha1.Kind) (*ResourceData, error) {
 	defer s.mu.RUnlock()
 	concreteResourceData, exists := s.registered[kind]
 	if !exists {
-		return &ResourceData{}, NewNotFound("TargetKind %s was not found", kind)
+		return &ResourceData{}, bindErr.NewNotFound("TargetKind %s was not found", kind)
 	}
 	return concreteResourceData, nil
 }
